+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to this project will be documented in this file. Changes not
 # [Unreleased]
 
 ## Added
-- Add new case to the responses (`noContent`)  by [@tomieq](https://github.com/tomieq)
+- Add new cases to the responses (`noContent`, `notImplemented`, `badGateway`, `serviceUnavailable`)  by [@tomieq](https://github.com/tomieq)
 - Replace dictionary headers with HttpResponseHeaders class  by [@tomieq](https://github.com/tomieq)
 - Keep in HttpRequest cookie values  by [@tomieq](https://github.com/tomieq)
 - Replace JSONSerialization with Encodable  by [@tomieq](https://github.com/tomieq)